requirements.txt
streamlit
google-generativeai
nltk
spacy
python-docx
